Trying to figure stuff out.
Most of these changes were just attempts at getting the data into the Master Table. I was having issues with every value appearing in a single column. I searched for a tutorial and found that and will follow it to see if that will get me where I need to go.