Table is working minus pictures
I have the table pulling in the data how I want so now it's about tweaking and seeing what data I want in it.

I am stuck on getting the headshots in the table however. Every time I try and use a template literal in my image src it breaks the table and it doesn't render on screen.

I have a Stack Overflow question that I will go over to see if I can get it working. Invovles a method.