Part 1 of MasterTable Front End is complete
The Master Table front end is essentially complete. I have all the data in my test JSON generating in the table with no issues. Now, it's about styalizing and adding any additional columns that would be appropriate. The next step is to add functionality to add the drafted players over to the DraftedList.