Master Table Front End v1.0.0
* Trying to figure stuff out.

Most of these changes were just attempts at getting the data into the Master Table. I was having issues with every value appearing in a single column. I searched for a tutorial and found that and will follow it to see if that will get me where I need to go.

* Table is working minus pictures

I have the table pulling in the data how I want so now it's about tweaking and seeing what data I want in it.

I am stuck on getting the headshots in the table however. Every time I try and use a template literal in my image src it breaks the table and it doesn't render on screen.

I have a Stack Overflow question that I will go over to see if I can get it working. Invovles a method.

* Nothing added

Just kept testing getting the image in the table but still don't have a fix.

* Part 1 of MasterTable Front End is complete

The Master Table front end is essentially complete. I have all the data in my test JSON generating in the table with no issues. Now, it's about styalizing and adding any additional columns that would be appropriate. The next step is to add functionality to add the drafted players over to the DraftedList.

* Started to build in sorting for the table

* Completed MasterTable Front End Build

The MasterTable Front End build is finsihed for now.

Here is a list of things it does:

- Brings in data consisting of player rank, position, position rank, headshot, name, team, bye, projected points, projected points per game, and a draft button that currently doesn't do anything

- I've added a next and previous page button so only 10 players show at a time. I will probably bump that up to 15-25 once I have full data in the app

- Each column can be sorted

I did make some extra styalizing changes but will come back at a later date to finish up design aspects later. I am wanting to get functionality in this app before getting into the nitty gritty of CSS.