Moving player data around
I was able to add functionality to push the player data of the button being selected into an array in the data option.

The goal is to be able to use that data and transfer it over to another component to be used.

I also was able to remove the row of the drafted player from the rendered list. The next step is going to add an undo button to bring back the most recent changes back into the list.